Give the error codes in c4Base.h explicit enum typedefs.
This helps the C# bindings-generator tool.

+++ C/c4Base.cc
@@ -40,7 +40,7 @@ namespace c4Internal {
 
     static_assert((int)kC4MaxErrorDomainPlus1 == (int)error::NumDomainsPlus1,
                   "C4 error domains are not in sync with C++ ones");
-    static_assert(kC4NumErrorCodesPlus1 == error::NumLiteCoreErrorsPlus1,
+    static_assert((int)kC4NumErrorCodesPlus1 == (int)error::NumLiteCoreErrorsPlus1,
                   "C4 error codes are not in sync with C++ ones");
 
     // A buffer that stores recently generated error messages, referenced by C4Error.internal_info
